Autonomous Launch Orchestrator/autonomous_launch_orchestrator/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langchain
langchain-community
langchain-google-genai
openai
python-dotenv
requests

